🌟 feat: 닉네임 변경 추가

+++ src/main/java/com/x1/groo/user/repository/UserRepository.java
@@ -4,7 +4,7 @@ import com.x1.groo.user.aggregate.UserEntity;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
     Optional<UserEntity> findByEmail(String email);
 
     boolean existsByEmail(String email);
